完成函数调用 Confirm 组件

+++ src/views/layout/components/header/header-search/history.vue
@@ -33,6 +33,7 @@ const EMITS_ITEM_CLICK = 'itemClick'
 </script>
 <script setup>
 import { useStore } from 'vuex'
+import { confirm } from '@/libs'
 
 const emits = defineEmits([EMITS_ITEM_CLICK])
 const store = useStore()
@@ -40,7 +41,11 @@ const store = useStore()
  * 删除全部
  */
 const onDeleteAllClick = () => {
-  store.commit('search/deleteAllHistory')
+  confirm('你要删除所有记录吗？')
+    .then(() => {
+      store.commit('search/deleteAllHistory')
+    })
+    .catch(() => {})
 }
 
 /**
